refactor: 패키지 이름 mockData로 수정
Related: #54
