+++ src/main/java/com/modoospace/mockData/controller/dto/address/Document.java
@@ -1,4 +1,4 @@
-package com.modoospace.data.controller.dto.space;
+package com.modoospace.mockData.controller.dto.address;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -8,10 +8,13 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class SpaceInfo {
+public class Document {
 
-    String name;
+    JibunAddress address;
 
-    @JsonProperty("desc")
-    String description;
+    @JsonProperty("road_address")
+    RoadAddress roadAddress;
+
+    String x;
+    String y;
 }
