+++ src/main/java/com/modoospace/mockData/controller/dto/space/BreakDay.java
@@ -1,4 +1,4 @@
-package com.modoospace.data.controller.dto.address;
+package com.modoospace.mockData.controller.dto.space;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -8,13 +8,11 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Document {
+public class BreakDay {
 
-    JibunAddress address;
+    @JsonProperty("BRK_DAY_TP_CD")
+    String breakDayCd;
 
-    @JsonProperty("road_address")
-    RoadAddress roadAddress;
-
-    String x;
-    String y;
+    @JsonProperty("DAYW_CD")
+    String dayCd;
 }
