+++ src/main/java/com/modoospace/mockData/controller/dto/space/BreakTime.java
@@ -1,4 +1,4 @@
-package com.modoospace.data.controller.dto.space;
+package com.modoospace.mockData.controller.dto.space;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -8,11 +8,11 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class BreakDay {
+public class BreakTime {
 
-    @JsonProperty("BRK_DAY_TP_CD")
-    String breakDayCd;
+    @JsonProperty("start_time")
+    Integer startHour;
 
-    @JsonProperty("DAYW_CD")
-    String dayCd;
+    @JsonProperty("end_time")
+    Integer endHour;
 }
