+++ src/main/java/com/modoospace/mockData/controller/dto/space/FacilityResponse.java
@@ -1,4 +1,4 @@
-package com.modoospace.data.controller.dto.space;
+package com.modoospace.mockData.controller.dto.space;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.modoospace.space.domain.Facility;
@@ -8,6 +8,7 @@ import com.modoospace.space.domain.WeekdaySettings;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Getter
@@ -16,7 +17,7 @@ import java.util.List;
 public class FacilityResponse {
 
     FacilityInfo info;
-    List<FacilityCategory> categories;
+    List<FacilityCategory> categories = new ArrayList<>();
 
     public Facility toFacility(TimeSettings timeSettings, WeekdaySettings weekdaySettings, Space space) {
         return Facility.builder()
