+++ src/main/java/com/modoospace/mockData/controller/dto/space/SpaceInfo.java
@@ -1,4 +1,4 @@
-package com.modoospace.data.controller.dto.space;
+package com.modoospace.mockData.controller.dto.space;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -8,11 +8,10 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class BreakHoliday {
+public class SpaceInfo {
 
-    @JsonProperty("BRK_DAY_TP_CD")
-    String breakDayCd;
+    String name;
 
-    @JsonProperty("DAYW_CD")
-    String dayCd;
+    @JsonProperty("desc")
+    String description;
 }
