+++ src/main/java/com/modoospace/space/domain/Space.java
@@ -93,8 +93,4 @@ public class Space extends BaseTimeEntity {
      private boolean hasFacilities() {
         return !facilities.isEmpty();
     }
-
-    public void setFacilities(List<Facility> facilities) {
-        this.facilities.addAll(facilities);
-    }
 }
